miniproject1/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
celery
redis
drf-yasg
django-filter
Pillow
django-crispy-forms
crispy-bootstrap4
pdfkit
